user role tabelle

+++ database/migrations/2024_11_17_140844_create_user_role.php
@@ -11,12 +11,13 @@ return new class extends Migration
      */
     public function up(): void
     {
-        Schema::create('user_role', function (Blueprint $table) {
-            $table->foreignId('user_id')->constrained()->onDelete('cascade');
-            $table->foreignId('role_id')->constrained()->onDelete('cascade');
-            $table->primary(['user_id', 'role_id']);
-        });
-
+        if (!Schema::hasTable('user_role')) {
+            Schema::create('user_role', function (Blueprint $table) {
+                $table->foreignId('user_id')->constrained()->onDelete('cascade');
+                $table->foreignId('role_id')->constrained()->onDelete('cascade');
+                $table->primary(['user_id', 'role_id']);
+            });
+        }
     }
 
     /**
